refactor(#155): move isEqualIntersection to utils
